requirements.txt
django
djangorestframework
djoser
mysqlclient
djangorestframework-xml